Refactor recipe embedding worker to pure queue consumer architecture
